Many stuff wow

+++ Logic/source/AI/Enemy.cpp
@@ -9,7 +9,7 @@
 
 using namespace Logic;
 
-Enemy::Enemy(Graphics::ModelID modelID, btRigidBody* body, btVector3 halfExtent, float health, float baseDamage, float moveSpeed, ENEMY_TYPE enemyType, int animationId)
+Enemy::Enemy(Graphics::ModelID modelID, btRigidBody* body, btVector3 halfExtent, int health, int baseDamage, float moveSpeed, ENEMY_TYPE enemyType, int animationId)
 : Entity(body, halfExtent, modelID)
 {
 	m_behavior = nullptr;
@@ -88,7 +88,7 @@ bool Enemy::hasCallbackEntities()
     return m_nrOfCallbacksEntities > 0;
 }
 
-void Enemy::damage(float damage)
+void Enemy::damage(int damage)
 {
 	m_health -= damage;
 
@@ -104,24 +104,24 @@ void Enemy::affect(int stacks, Effect const &effect, float dt)
 	if (flags & Effect::EFFECT_KILL)
 		damage(m_health);
 	if (flags & Effect::EFFECT_ON_FIRE)
-		damage(effect.getModifiers()->modifyDmgTaken * dt);
+		damage(static_cast<int> (effect.getModifiers()->modifyDmgTaken * dt));
 	if (flags & Effect::EFFECT_BULLET_TIME)
 		m_bulletTimeMod *= std::pow(effect.getSpecifics()->isBulletTime, stacks);
     if (flags & Effect::EFFECT_IS_FROZEN)
         m_moveSpeedMod *= std::pow(effect.getSpecifics()->isFreezing, stacks);
 }
 
-float Enemy::getHealth() const
+int Enemy::getHealth() const
 {
 	return m_health;
 }
 
-float Enemy::getMaxHealth() const
+int Enemy::getMaxHealth() const
 {
 	return m_maxHealth;
 }
 
-float Enemy::getBaseDamage() const
+int Enemy::getBaseDamage() const
 {
 	return m_baseDamage;
 }
